--- hobby_games/db.py
@@ -1,12 +1,0 @@
-from sqlalchemy import create_engine
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import scoped_session, sessionmaker
-import os, sys
-sys.path.append(os.path.dirname(os.path.dirname(os.path.realpath(__file__))))
-from webapp.config import DB
-
-engine = create_engine(DB)
-db_session = scoped_session(sessionmaker(bind=engine))
-
-Base = declarative_base()
-Base.query = db_session.query_property()
